exercises/04-01-Filtering-server/requirements.txt
pandas
seaborn
paho-mqtt
aiocoap
